ascs: Fix possible NULL pointer dereference
This fixes exception that might happen when the CIS is disconnected
locally as the one of the steps performed in `Releasing` state.
The function that is called then, bt_audio_stream_detach sets ep->stream
to NULL and when we receive HCI Disconnected event with CIS handle (the
one for witch we already cleaned up endpoint) we hit this NULL pointer
dereference.

Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>