Bluetooth: audio: tbs_client: Handle peer being unpaired
Once peer is unpaired, the bt_gatt_subscribe_params.notify
callback is called with conn == NULL to inform that the subscription has
been removed.
This fixes an assert that was fired when conn was NULL.
The notification callback points to the actual handler, so
CONTAINER_OF macro can be used to find the TBS instance.
This reduced the code complexity as redundant conditionals have been
removed.

Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>